requirements.txt
tk
matplotlib
numpy<2
opencv-python
scikit-learn
tensorboard
torch
torchvision
PyYAML
